refactor(index builder): Move common logic to abstract list builder
US-32041
