+++ src/Redis/Indexes/Builders/ListBuilder.php
@@ -18,122 +18,8 @@ use Predis\Response\ServerException;
 use Predis\Transaction\AbortedMultiExecException;
 use Predis\Transaction\MultiExec;
 
-class ListBuilder
+class ListBuilder extends AbstractListBuilder
 {
-    /**
-     * Redis
-     *
-     * @var Client|ClientInterface
-     */
-    protected $redis;
-
-    /**
-     * Keys resolver
-     *
-     * @var KeysResolver
-     */
-    protected $keys;
-
-    /**
-     * ListBuilder constructor.
-     *
-     * @param ClientInterface $redis Redis client
-     * @param KeysResolver    $keys  Keys
-     */
-    public function __construct(ClientInterface $redis, KeysResolver $keys)
-    {
-        $this->redis = $redis;
-        $this->keys  = $keys;
-    }
-
-    /**
-     * Building index with new state
-     *
-     * @param ListDto $dto   List dto
-     * @param State   $state State
-     *
-     * @return void
-     */
-    public function build(ListDto $dto, State $state) : void
-    {
-        if ($state->isShouldBeFlushed() === true) {
-            $this->flush($dto);
-        } else if ($state->hasDeletedItems() === true) {
-            $this->delete($dto, $state->getDeletedItems());
-        }
-
-        $this->add($dto, $state->getAddedItems());
-    }
-
-    /**
-     * Start watching for changes in index
-     *
-     * @param ListDto $dto List dto
-     *
-     * @return void
-     */
-    protected function watch(ListDto $dto) : void
-    {
-        $watchKey = $this->keys->makeWatchingOnDocIndexKey($dto->docType, $dto->docId, $dto->name);
-        $this->redis->watch($watchKey);
-    }
-
-    /**
-     * Begin transaction
-     *
-     * @param ListDto $dto List dto
-     *
-     * @return MultiExec
-     */
-    protected function beginTransaction(ListDto $dto) : MultiExec
-    {
-        $transaction = $this->redis->transaction();
-        $transaction->setex($this->keys->makeWatchingOnDocIndexKey($dto->docType, $dto->docId, $dto->name), 1, "");
-
-        return $transaction;
-    }
-
-    /**
-     * Getting system key of hash where we stored actual states of indexes
-     *
-     * @param ListDto $dto List dto
-     *
-     * @return string
-     */
-    protected function getSysKey(ListDto $dto) : string
-    {
-        return $this->keys->makeIndexSysKey($dto->docType);
-    }
-
-    /**
-     * Getting field of hash where we stored actual state
-     *
-     * @param ListDto $dto List dto
-     *
-     * @return string
-     */
-    protected function getSysField(ListDto $dto) : string
-    {
-        return $this->keys->makeSysField($dto->docId, $dto->docType);
-    }
-
-    /**
-     * Getting actual state
-     *
-     * @param ListDto $dto List dto
-     *
-     * @return array
-     */
-    protected function getActualState(ListDto $dto) : array
-    {
-        $data = $this->redis->hget($this->getSysKey($dto), $this->getSysField($dto));
-
-        if (empty($data) === true) {
-            return [];
-        }
-
-        return (array) json_decode($data, true);
-    }
 
     /**
      * Flush actual state, be careful
